2. Body posture detection/OpenPose demo/requirements.txt
ffmpeg-python>=0.2.0
numpy==1.26.4          # 与 Matplotlib 3.7.x ABI 匹配
matplotlib==3.7.3
torch==2.5.1+cu121 ; sys_platform == "win32"
torchvision==0.20.1+cu121 ; sys_platform == "win32"
torchaudio==2.5.1+cu121  ; sys_platform == "win32"
--extra-index-url https://download.pytorch.org/whl/cu121
opencv-python>=4.11.0.86
